+++ python/packages/autogen-ext/src/autogen_ext/agents/video_surfer/__init__.py
@@ -1,19 +1,7 @@
-from ._action_space import (
-    extract_audio,
-    get_screenshot_at,
-    get_video_length,
-    openai_transcribe_video_screenshot,
-    save_screenshot,
-    transcribe_audio_with_timestamps,
-)
+from . import tools
 from ._video_surfer import VideoSurferAgent
 
 __all__ = [
     "VideoSurferAgent",
-    "extract_audio",
-    "transcribe_audio_with_timestamps",
-    "get_video_length",
-    "save_screenshot",
-    "openai_transcribe_video_screenshot",
-    "get_screenshot_at",
+    "tools",  # Expose the tools submodule
 ]
