+++ python/packages/autogen-studio/autogenstudio/database/component_factory.py
@@ -10,7 +10,7 @@ from autogen_agentchat.agents import AssistantAgent, UserProxyAgent
 from autogen_agentchat.task import MaxMessageTermination, StopMessageTermination, TextMentionTermination
 from autogen_agentchat.teams import RoundRobinGroupChat, SelectorGroupChat
 from autogen_core.components.tools import FunctionTool
-from autogen_ext.agents import MultimodalWebSurfer
+from autogen_ext.agents.web_surfer import MultimodalWebSurfer
 from autogen_ext.models import OpenAIChatCompletionClient
 
 from ..datamodel.types import (
